project3-rag-system/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic>=2.7.4
python-dotenv==1.2.2
psycopg2-binary==2.9.10
python-multipart==0.0.20
google-genai==1.68.0
langchain==1.2.13
langchain-google-genai==4.2.1
langchain-text-splitters==1.1.1
pytest==9.0.2
